refactor: uncouple task component from tasklist
