+++ src/components/Task/Task.d.ts
@@ -1,0 +1,6 @@
+import { Tarefa } from '../../utils/model';
+
+export type TaskProps = {
+    task: Tarefa;
+    onTaskChange: (taskId: number) => void;
+}
