+++ src/components/TaskList/TaskList.d.ts
@@ -7,4 +7,5 @@ export type TaskListWrapperProps = {
 
 export type TaskListProps = {
     tasks: Tarefa[]
+    categoria: Categoria;
 }
